move iconwrapper to the input
